% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cohRel.R
\name{compute_cohRel}
\alias{compute_cohRel}
\title{Calculates the response coherence and reliability}
\usage{
compute_cohRel(df, scales_list, nb_factors, max_iterations = 30)
}
\arguments{
\item{df}{a dataframe of data (e.g. survey responses); a matrix won't work}

\item{scales_list}{a list of length number of scale indicating the name of the scales
included in the df and the items for each scale.}

\item{nb_factors}{a list of length number of scales indicating the name of each scale
and the number of latent factors for each scale included in the df}

\item{max_iterations}{a numerical value indicating the maximal number of iterations
allowed to obtain strict orthogonality and perfect communality, assuming
the model does not converge before reaching said number of iterations. This allows you to
control the potential computational load. Recommended value = 30}
}
\description{
Takes a dataframe of item responses, a list specifying the scale names and
the items belonging to it, and a list indicating the number of latent factors for each scale.
The response reliability and coherence is then obtained by running a series of PCA, computing
the vector of response strategy and by computing the norm of the response strategy (response coherence)
and the correlation between two parallel halves  of the response strategy. The function returns a list
which contains 2 sub-list of length number of scales. Depending on the number of latent factors in your scale,
it might take from a few seconds to a few minutes for the code to run.
}
\examples{
df <- mierda_data #rename dataframe
scales_list <- list(scale1 = dplyr::select (df, Item_1:Item_50))
nb_factors <- list(scale1 = 5)
rel_coh <- compute_cohRel(df, scales_list, nb_factors)
}
\references{
Dupuis, M., Meier, E., Capel, R., & Gendre, F. (2015). Measuring individuals’ response quality
in self-administered psychological tests: An introduction to Gendre’s functional method.
\emph{Frontiers in Psychology, 6}, 629. \doi{10.3389/FPSYG.2015.00629}
}
\author{
Ariane J. Gauthier \email{arianejgauthier@outlook.com}
}
