% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_isd.R
\name{compute_isd}
\alias{compute_isd}
\title{Calculates the intra-individual variability (IRV) for all the rows of a given dataset}
\usage{
compute_isd(df, na.rm = TRUE)
}
\arguments{
\item{df}{a dataframe of matrix of data (e.g. survey responses)}
}
\description{
Takes a dataframe or matrix of item responses and returns for each participant
the within-person standard deviation
}
\examples{
isd <- compute_isd(df, na.rm = TRUE)

}
\references{
Dunn, A. M., Heggestad, E. D., Shanock, L. R., & Theilgard, N. (2018).
Intra-individual Response Variability as an Indicator of Insufficient Effort Responding:
Comparison to Other Indicators and Relationships with Individual Differences.
\emph{Journal of Business and Psychology, 33(1)}, 105-121. \doi{10.1007/s10869-016-9479-0}

Marjanovic, Z., Holden, R., Struthers, W., Cribbie, R., & Greenglass, E. (2015).
The inter-item standard deviation (ISD): An index that discriminates between conscientious and random responders.
\emph{Personality and Individual Differences}, 84, 79-83. \doi{10.1016/j.paid.2014.08.021}
}
\author{
Ariane J. Gauthier \email{arianejgauthier@outlook.com}
}
