% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_longstring.R
\name{compute_longstring}
\alias{compute_longstring}
\title{Calculates the longest string of repetitive responses in a specified dataset}
\usage{
compute_longstring(df)
}
\arguments{
\item{df}{a dataframe of matrix of data (e.g. survey responses)}
}
\description{
Takes a dataframe or matrix of item responses and returns for each participant
the lenght of the longest repetitive string of consecutive responses.
}
\examples{
df <- mierda_data
ls <- compute_longstring(df)

}
\references{
Johnson, J. A. (2005). Ascertaining the validity of individual protocols
from web-based personality inventories. \emph{Journal of Research in Personality, 39}, 103-129. \doi{10.1016/j.jrp.2004.09.009}
}
\author{
Ariane J. Gauthier \email{arianejgauthier@outlook.com}
}
