% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_md.R
\name{compute_md}
\alias{compute_md}
\title{Calculates the Mahalanobis Distance for the participants within a dataset}
\usage{
compute_md(df, return_pvalues = FALSE)
}
\arguments{
\item{df}{is a dataframe or matrix containing respondents responses (e.g. survey responses)}
}
\description{
akes a dataframe or matrix of item responses, and applies Mahalanobis Distance.
Returns the distance and associated p-value for each respondent.
   IF return_pvalues = FALSE, returns a vector containing the distance;
   IF return_pvalues = TRUE returns a list containing the distance and p-value.
}
\examples{
md <- compute_md(df, return_pvalues = FALSE)
md <- compute_md(df, return_pvalues = TRUE)

}
\references{
#' Meade, A. W., & Craig, S. B. (2012). Identifying careless responses in survey data.
\emph{Psychological Methods, 17(3)}, 437-455. \doi{10.1037/a0028085}
}
\author{
Ariane J. Gauthier \email{arianejgauthier@outlook.com}
}
