% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_ptc.R
\name{compute_ptc}
\alias{compute_ptc}
\title{Calculates the person-total correlation in a specified dataset}
\usage{
compute_ptc(df)
}
\arguments{
\item{df}{a dataframe of matrix of data (e.g. survey responses)}
}
\description{
Takes a dataframe or matrix of item responses and returns for each participant
the correlation between their responses and the mean responses observed at a sample-level
}
\examples{
ptc <- compute_ptc(df)

}
\references{
Karabatsos, G. (2003). Comparing the aberrant response detection performance of
thirty six person-fit statistics. \emph{Applied Measurement in Education, 16}, 277–298. \doi{10.1207/S15324818AME1604_2}
}
\author{
Ariane J. Gauthier \email{arianejgauthier@outlook.com}
}
