% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mIERda_cutoff.R
\name{mIERda_cutoff}
\alias{mIERda_cutoff}
\alias{mierda_cutoff}
\title{Computes the data-driven cutoff based on the mIERda scores}
\usage{
mIERda_cutoff(score)
}
\arguments{
\item{score}{a vector containing respondents mierda score returned by the mierda_score() function}
}
\description{
The mIERda_cutoff function takes the vector of mIERda scores and returns a numeric value representing the optimal cutoff.
}
\author{
Ariane J. Gauthier \email{arianejgauthier@outlook.com}
}
