% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mierda_data.R
\docType{data}
\name{mierda_data}
\alias{mierda_data}
\title{Simulated dataset with insufficient effort responding observations}
\format{
A dataframe with 500 observations (rows) and 50 variables (columns).
}
\usage{
mierda_data
}
\description{
The dataset was generated to recreate the presence of insufficient effort responding/careless observations.
IER observations were simulated either under a normal distribution, uniform distribution, bimodal distribution, moderately invariant, and highly invariant response pattern
Attentive observations were simulated from an IRT model and to create responses to the Big Five Inventory-50
Data simulated 5 subscales of 10 items each (= 50 variables).
Data contains 30% IER observations.
}
\keyword{datasets}
