% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mIERda_score.R
\name{mIERda_score}
\alias{mIERda_score}
\alias{mierda_score}
\title{Computes the mIERda score}
\usage{
mIERda_score(x)
}
\arguments{
\item{x}{a dataframe of matrix containing respondents values on multiple IERIs}
}
\description{
The mIERda_score function takes a dataframe or matrix containing participants ISD, PTC, longtring,
response reliability and response coherence and generates a fair cut forest. A vector containing mIERda scores is returned.
}
\examples{
x <- cbind(ls, md, ptc, isd, coh_rel, resp_rel)
score <- mIERda_score(x)

}
\references{
\enumerate{
\item Liu, Fei Tony, Kai Ming Ting, and Zhi-Hua Zhou. "Isolation forest." 2008 Eighth IEEE International Conference on Data Mining. IEEE, 2008.
\item Liu, Fei Tony, Kai Ming Ting, and Zhi-Hua Zhou. "Isolation-based anomaly detection." ACM Transactions on Knowledge Discovery from Data (TKDD) 6.1 (2012): 3.
\item Cortes, David. "Imputing missing values with unsupervised random trees." arXiv preprint arXiv:1911.06646 (2019).
\item Cortes, David. "Revisiting randomized choices in isolation forests." arXiv preprint arXiv:2110.13402 (2021).
\item Cortes, David. "Isolation forests: looking beyond tree depth." arXiv preprint arXiv:2111.11639 (2021).
}
}
\author{
Ariane J. Gauthier \email{arianejgauthier@outlook.com}
}
