% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cohRel.R
\name{compute_cohRel}
\alias{compute_cohRel}
\title{Calculates the response coherence and reliability}
\usage{
compute_cohRel(df, scales_list, nb_factors)
}
\arguments{
\item{df}{a dataframe of data (e.g. survey responses); a matrix won't work}

\item{scales_list}{a list of length number of scale indicating the name of the scales
included in the df and the items for each scale.}

\item{nb_factors}{a list of length number of scales indicating the name of each scale
and the number of latent factors for each scale included in the df}
}
\description{
Takes a dataframe of item responses, a list specifying the scale names and
the items belonging to it, and a list indicating the number of latent factors for each scale.
The response reliability and coherence is then obtained by running a series of PCA, computing
the vector of response strategy and by computing the norm of the response strategy (response coherence)
and the correlation between two parallel halves  of the response strategy. The function returns a list
which contains 2 sub-list of length number of scales.
}
\examples{
 scales_list <- list(scale1 = dplyr::select (df, item_1:item_18),
                scale2 = dplyr::select (df, item_19:item_29),
                scale3 = dplyr::select (df, item_30:item_60))

 nb_factors <- list(scale1 = 6,
                    scale2 = 2,
                    scale3 = 2)

response_coh_rel <- response_cohRel(df, scales_list, nb_factors)
}
\references{
Dupuis, M., Meier, E., Capel, R., & Gendre, F. (2015). Measuring individuals’ response quality
in self-administered psychological tests: An introduction to Gendre’s functional method.
\emph{Frontiers in Psychology, 6}, 629. \doi{10.3389/FPSYG.2015.00629}
}
\author{
Ariane J. Gauthier \email{arianejgauthier@outlook.com}
}
