% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mIERda_cutoff.R
\name{mIERda_cutoff}
\alias{mIERda_cutoff}
\title{Computes the data-driven cutoff based on the mIERda scores}
\usage{
mIERda_cutoff(mierda_score)
}
\arguments{
\item{x}{a vector containung respondents mierda_score}
}
\description{
The mIERda_cutoff function takes the vector of mIERda scores and returns a numeric value representing the optimal cutoff.
}
\examples{
cutoff <- mIERda_cutoff(mierdaScore)


}
\author{
Ariane J. Gauthier \email{arianejgauthier@outlook.com}
}
